+++ feature/CourseView/Sources/CourseView/Config.swift
@@ -1,4 +1,4 @@
 // swiftlint:disable identifier_name
 
-let MIN_SCORE_GREEN: Float = 80
-let MIN_SCORE_ORANGE: Float = 40
+let MIN_SCORE_GREEN: Double = 80
+let MIN_SCORE_ORANGE: Double = 40
